+++ test/cases/migration/create_join_table_test.rb
@@ -21,54 +21,6 @@ module ActiveRecord
           end
         end
 
-        def test_create_join_table
-          connection.create_join_table :artists, :musics
-
-          assert_equal %w(artist_id music_id rowid), connection.columns(:artists_musics).map(&:name).sort
-        end
-
-        def test_create_join_table_set_not_null_by_default
-          connection.create_join_table :artists, :musics
-
-          assert_equal [false, false, false], connection.columns(:artists_musics).map(&:null)
-        end
-
-        def test_create_join_table_with_strings
-          connection.create_join_table "artists", "musics"
-
-          assert_equal %w(artist_id music_id rowid), connection.columns(:artists_musics).map(&:name).sort
-        end
-
-        def test_create_join_table_with_symbol_and_string
-          connection.create_join_table :artists, "musics"
-
-          assert_equal %w(artist_id music_id rowid), connection.columns(:artists_musics).map(&:name).sort
-        end
-
-        def test_create_join_table_with_the_proper_order
-          connection.create_join_table :videos, :musics
-
-          assert_equal %w(music_id rowid video_id), connection.columns(:musics_videos).map(&:name).sort
-        end
-
-        def test_create_join_table_with_the_table_name
-          connection.create_join_table :artists, :musics, table_name: :catalog
-
-          assert_equal %w(artist_id music_id rowid), connection.columns(:catalog).map(&:name).sort
-        end
-
-        def test_create_join_table_with_the_table_name_as_string
-          connection.create_join_table :artists, :musics, table_name: "catalog"
-
-          assert_equal %w(artist_id music_id rowid), connection.columns(:catalog).map(&:name).sort
-        end
-
-        def test_create_join_table_with_column_options
-          connection.create_join_table :artists, :musics, column_options: { null: true }
-
-          assert_equal [true, true, false], connection.columns(:artists_musics).map(&:null)
-        end
-
         # This test is identical to the one found in Rails, save for the fact
         # that transactions are turned off for test runs. It is necessary to disable
         # transactional tests in order to assert on schema changes due to how
@@ -93,11 +45,6 @@ module ActiveRecord
             connection.drop_join_table "audio_artists", "audio_musics"
             assert !connection.table_exists?("audio_artists_musics"), "Should have dropped join table, but didn't"
           end
-        end
-
-        def test_create_join_table_with_uuid
-          connection.create_join_table :artists, :musics, column_options: { type: :uuid }
-          assert_equal [:uuid, :uuid, :integer], connection.columns(:artists_musics).map(&:type)
         end
 
         private